Upstream changes from Zakodium (#36)
* feat: update BullMQ to v5

BREAKING-CHANGE: New BullMQ version changes the data format and requires to upgrade all workers to the new version.

* fix: better error message when job cannot be loaded

Not found is not always the reason. For example, loading a job's dependency could throw.
Logging the error helps to understand the underlying issue.

* fix: typings for clear, list and get methods

- `list` and `get` don't return promises
- The generic type of `clear` is useless

* feat: add `getOrSet` method

Like `get`, but instantiates the queue if it doesn't exist yet.
Without this method, the only way to get a queue other than the default one is to dispatch a job first.

* fix: close queues when Adonis app is shutdown

Otherwise it stays alive forever because of the open Redis connection(s).

* fix: do not create default queue on instantiation

This avoids creating a connection to Redis when it's not needed, for example in Ace commands or unit tests.